Create an abstract class for the matcher
